+++ R/growthModelSim.R
@@ -249,7 +249,7 @@ iPredLength <- function(type,t,p1,p2,p3,p4) {
          # p1=omega,p2=K,  p3=t0, p4 not used
          vbGQ=, vbGallucciQuinn= { sd <- (p1/p2)*(1-exp(-p2*(t-p3))) },
          # p1=Linf, p2=K0,  p3=to, p4 not used
-         vbWeisberg= {  sd <- p1*(1-exp(-(log(2)/p2)*(t-p3))) },
+         vbWeisberg= {  sd <- p1*(1-exp(-(log(2)/(p2-p3))*(t-p3))) },
          # p1=Linf, p2=L0, p3=ome,p4 not used
          vbMooij= { sd <- p1-(p1-p2)*exp(-(p3/p1)*t) },
          # p1=L1,   p2=L2, p3=K,  p4=b